CODE OVERHAUL!!!
